# Add split-kv and M<->H swap to varlen forward decoding attention