This PR speeds up decoding attention when lengths are variable within the batch, by adding two optimizations (split-kv and M<->H swap), which were already present on non-variable-length path. 

It also refactors logic for setting split-kv-related parameters into a separate function `set_params_splitkv`.

# Tests

```
pytest -k test_flash_attn_varlen tests/test_flash_attn.py 2>&1 | tee log.txt

platform linux -- Python 3.9.18, pytest-7.4.4, pluggy-1.3.0
rootdir: /home/grigorysizov/flash-attention
collected 367460 items / 312356 deselected / 55104 selected

tests/test_flash_attn.py ............................................... [  0%]
...
.................................................                        [100%]
============= 55104 passed, 312356 deselected in 771.03s (0:12:51) =============
```

# Benchmark

I've tweaked `benchmarks/benchmark_causal.py` for decoding setup with multiquery ([code](https://gist.github.com/sgrigory/1e970645cc4ffdfb9a61973b8b001ecf)). Here are the numbers for forward pass on H100:

On this PR:
```
batch_size=1, seqlen=2048: 0.047ms
batch_size=16, seqlen=2048: 0.046ms
batch_size=128, seqlen=2048: 0.073ms
batch_size=1, seqlen=4096: 0.046ms
batch_size=16, seqlen=4096: 0.046ms
batch_size=128, seqlen=4096: 0.135ms
batch_size=1, seqlen=8192: 0.037ms
batch_size=16, seqlen=8192: 0.042ms
batch_size=128, seqlen=8192: 0.265ms
```

On main:

```
batch_size=1, seqlen=2048: 0.059ms
batch_size=16, seqlen=2048: 0.099ms
batch_size=128, seqlen=2048: 0.668ms
batch_size=1, seqlen=4096: 0.124ms
batch_size=16, seqlen=4096: 0.231ms
batch_size=128, seqlen=4096: 1.387ms
batch_size=1, seqlen=8192: 0.249ms
batch_size=16, seqlen=8192: 0.430ms
batch_size=128, seqlen=8192: 2.849ms
```

You can see that speedup is ~10x for large batch sizes and seqlens. For small inputs the relative speedup is smaller.